+++ src/schemas/versionedQuestion.ts
@@ -48,6 +48,10 @@ type VersionedQuestion {
 
     "The conditional logic associated with this VersionedQuestion"
     versionedQuestionConditions: [VersionedQuestionCondition!]
+    "Owner affiliation for the question"
+    ownerAffiliation: Affiliation
+    "Versioned Section Tags associated with the question"
+    sectionTags: [Tag!]
 }
 
 "A snapshot of a Question when it became published, but includes extra information about if answer is filled."